Slopebug fix (#75)
* Add a cvar to fix slopebug

Adds the cl_slopebug_fix cvar, set to 0 by default. Set it to 1 when playing on AG 6.7 servers to match serverside physics and have a smooth surfing experience.

* Apply slopebug fix based on server message

AG 6.7 sends a byte in the InitHUD message, telling that it has slopefix. AG 6.6 and lower won't send this byte, but they do send the message, so the lack of that byte means it doesn't have slopefix.
InitHUD is sent only once, when you join a server. We're using this message because it allows us to apply the slopebug fix at the right time.
It's set automatically because the player may not know the server version, and setting it manually by trying if surf prediction is weird is a bit tedious.

* Simplify slopebug fix and correct code style

Co-authored-by: Ivan Molodetskikh <yalterz@gmail.com>

Co-authored-by: Ivan Molodetskikh <yalterz@gmail.com>